User and Admin Statements
Separated into two folders are the User and Admin calls that will be referenced in Python once we get to the coding stage. This may be altered later. This will save a lot of time in coding, as all we need to do is pull the individual code block in.